+++ vehicle/src/Vehicle/Compile/ObjectFile.hs
@@ -1,7 +1,6 @@
-
-module Vehicle.Compile.InterfaceFile
-  ( readInterfaceFile
-  , writeInterfaceFile
+module Vehicle.Compile.ObjectFile
+  ( readObjectFile
+  , writeObjectFile
   ) where
 
 import Control.Monad.IO.Class
@@ -15,24 +14,24 @@ import Vehicle.Compile.Prelude
 import Vehicle.Compile.Type (TypeCheckingResult(..))
 import Data.Hashable (Hashable(..))
 
-data InterfaceContents = InterfaceContents
+data ObjectFileContents = ObjectFileContents
   { _fileHash   :: Int
   , _typeResult :: TypeCheckingResult
   } deriving (Generic)
 
-instance ToJSON   InterfaceContents
-instance FromJSON InterfaceContents
+instance ToJSON   ObjectFileContents
+instance FromJSON ObjectFileContents
 
-getInterfaceFileFromSpecificationFile :: FilePath -> FilePath
-getInterfaceFileFromSpecificationFile specFile =
-  dropExtension specFile <> vehicleInterfaceFileExtension
+getObjectFileFromSpecificationFile :: FilePath -> FilePath
+getObjectFileFromSpecificationFile specFile =
+  dropExtension specFile <> vehicleObjectFileExtension
 
-readInterfaceFile :: (MonadLogger m, MonadIO m)
+readObjectFile :: (MonadLogger m, MonadIO m)
                   => FilePath
                   -> SpecificationText
                   -> m (Maybe TypeCheckingResult)
-readInterfaceFile specificationFile spec = do
-  let interfaceFile = getInterfaceFileFromSpecificationFile specificationFile
+readObjectFile specificationFile spec = do
+  let interfaceFile = getObjectFileFromSpecificationFile specificationFile
   errorOrContents <- liftIO $ do
     (Right <$> BIO.readFile interfaceFile) `catch` \ (e :: IOException) -> return (Left e)
 
@@ -46,7 +45,7 @@ readInterfaceFile specificationFile spec = do
         logDebug MinDetail $ "Unable to restore found interface file for" <+> quotePretty specificationFile
         return Nothing
 
-      Just (InterfaceContents specHash result)
+      Just (ObjectFileContents specHash result)
         | specHash /= hash spec -> do
           logDebug MinDetail $ "Outdated interface file found for" <+> quotePretty specificationFile
           return Nothing
@@ -55,13 +54,13 @@ readInterfaceFile specificationFile spec = do
           logDebug MinDetail $ "Loaded interface file for" <+> quotePretty specificationFile
           return $ Just result
 
-writeInterfaceFile :: MonadIO m
+writeObjectFile :: MonadIO m
                    => FilePath
                    -> SpecificationText
                    -> TypeCheckingResult
                    -> m ()
-writeInterfaceFile specificationFile spec result = do
-  let interfaceFile = getInterfaceFileFromSpecificationFile specificationFile
+writeObjectFile specificationFile spec result = do
+  let interfaceFile = getObjectFileFromSpecificationFile specificationFile
   let specHash = hash spec
-  let contents = InterfaceContents specHash result
+  let contents = ObjectFileContents specHash result
   liftIO $ BIO.writeFile interfaceFile (encode contents)